Changes as per review
