+++ api/taskcluster/webhook.go
@@ -331,9 +331,7 @@ func createAllRuns(
 				labelsForRun = append(labelsForRun, lastBit)
 			}
 
-			slowClient, cancel := aeAPI.GetSlowHTTPClient(resultsReceiverTimeout)
-			defer cancel()
-			uploadClient := uc.NewClient(slowClient, aeAPI)
+			uploadClient := uc.NewClient(aeAPI)
 			err := uploadClient.CreateRun(sha, username, password, urls, labelsForRun)
 			if err != nil {
 				errors <- err
